requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
requests
python-dotenv
psycopg2-binary